fastapi/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-dotenv==1.0.0
cohere==4.26.1
googletrans==4.0.0-rc1